Create README.md
I am 46 and I graduated from The Faculty of Philosophy in Novi Sad, Serbia, 
Department of English Language and Culture, on June 20th,2000 and I worked as an English 
language teacher in Sremska Kamenica Police High School until 2008 when I moved to Kruševac 
and from then on I have been working as a police analyst in Police Directorate of Kruševac. 
 
I have wealth of experience as a translator from Serbian to English language and from English to 
Serbian language as well as a teacher of English language to Serbian children and adults and I think 
that I would be a great asset to your company and that it will also give me the further experience 
of a team leader. 
 
I successfully completed OSCE Trainer Development Course / Mentoring Phase and I 
demonstrated trainer competence in accordance with OSCE standards as well as I successfully 
completed The Specialized Course on Written and Legal Translation in The Master Translation 
Agency in Belgrade, I passed TEFL Certificate of Completion, Teaching English Online 
Cambridge Assessment English Certificate and Teaching English Phonics in Early Childhood 
Certificate 

Also, I am working for an online Japanese company `Native Camp` but I am not satisfied with the 
salary which is around $1,7 for a 25-minute class. 
 
In my opinion: 
Teaching is more than a job, it’s a part of a teacher`s personality, and it could be fun: 
• through fun worksheets, fun teaching games, 
• through the fact that you get to learn new skills all the time, 
• There are lots of people to meet and relationships to develop and I love developing positive 
relationships, 
• You get to be a role model, influencing students for the good. You can help your students 
explore the world around them and open their eyes, 
• It’s awesome when students keep in touch, so you get to see how they’re using the skills 
they have learned from you, 
• It gives you the chance to be creative, to try new ideas, 
• It is learner-centered, 
• It offers you a chance to continuously get better, 
• You can teach to inspire the students, 
• It offers flexibility for both student and a teacher who can choose a mutually convenient 
time to meet online while not having to leave their homes, 
• Through promoting life-long learning, self-discipline, and connecting all to the global 
village. 
 
While teaching, I try to spark each student imagination, to find a hook in their heart and mind so 
that they feel a need to learn the material and I try to show them I do care about them as I am aware 
that as a teacher I am making a difference in their lives every day. They may forget what the teacher 
said but they will not forget how the teacher made them that day. I try to teach to inspire. I am 
married and have a nine-year-old son, I enjoy traveling, reading, and meeting new people. 
 
I may enclose the following diplomas and certificates of mine: 
 
• CEPOL Certificate 
• English Language Teacher Diploma 
• Legal Interpreter Diploma 
• OSCE Certificate 
• TEFL Certificate of Completion 
• Teaching English Online Cambridge Assessment English Certificate 
• Teaching English Phonics in Early Childhood Quensland University of Technology 
Certificate